+++ changelog.md
@@ -14,6 +14,7 @@
 ## [2.2.6] - 2016-MM-DD
 
 ### Added
+- --no-subs improve
 - Verify if subtitle files exists before download
 
 ## [2.2.5] - 2016-06-20