+++ changelog.md
@@ -6,18 +6,19 @@
 
 ### Add
 - Add --magnet argument
-- Add --episode argument
 - Clean cache
 - Save settings
 
 ## [2.2.6] - 2016-MM-DD
 
 ### Added
+- --episode argument
 - --season argument
 - --no-subs improve
 - Verify if subtitle files exists before download
 
 ### Changes
+- If only one show is found download with ask
 - If only one subtitles is found download with ask
 
 ## [2.2.5] - 2016-06-20